PORTALS-3578
Lazy-load routes in AMP-ALS portal
