--- apps/portals/ampals/src/config/explorePageRoutes.tsx
@@ -1,31 +1,0 @@
-import { RouteObject } from 'react-router'
-import { CardContainerLogic, QueryWrapperPlotNav } from 'synapse-react-client'
-import { datasetCollectionsCardContainerLogicProps } from './synapseConfigs/datasetcollections'
-import { datasetQueryWrapperPlotNavProps } from './synapseConfigs/datasets'
-import { programsCardContainerLogicProps } from './synapseConfigs/programs'
-import { filesQueryWrapperPlotNavProps } from './synapseConfigs'
-
-const explorePageRoutes: RouteObject[] = [
-  {
-    path: 'Programs',
-    element: <CardContainerLogic {...programsCardContainerLogicProps} />,
-  },
-  {
-    path: 'Collections',
-    element: (
-      <div className={'CollectionList'}>
-        <CardContainerLogic {...datasetCollectionsCardContainerLogicProps} />
-      </div>
-    ),
-  },
-  {
-    path: 'Datasets',
-    element: <QueryWrapperPlotNav {...datasetQueryWrapperPlotNavProps} />,
-  },
-  {
-    path: 'Files',
-    element: <QueryWrapperPlotNav {...filesQueryWrapperPlotNavProps} />,
-  },
-]
-
-export default explorePageRoutes
