+++ apps/portals/ampals/src/pages/Explore/datasets.tsx
@@ -1,0 +1,8 @@
+import { datasetQueryWrapperPlotNavProps } from '@/config/synapseConfigs/datasets'
+import { QueryWrapperPlotNav } from 'synapse-react-client/components/QueryWrapperPlotNav'
+
+function ExploreDatasets() {
+  return <QueryWrapperPlotNav {...datasetQueryWrapperPlotNavProps} />
+}
+
+export default ExploreDatasets
