+++ apps/portals/ampals/src/pages/Explore/files.tsx
@@ -1,0 +1,8 @@
+import { QueryWrapperPlotNav } from 'synapse-react-client/components/QueryWrapperPlotNav'
+import { filesQueryWrapperPlotNavProps } from '@/config/synapseConfigs/data'
+
+function ExploreFiles() {
+  return <QueryWrapperPlotNav {...filesQueryWrapperPlotNavProps} />
+}
+
+export default ExploreFiles
