+++ apps/portals/ampals/src/pages/Explore/programs.tsx
@@ -1,0 +1,8 @@
+import { programsCardContainerLogicProps } from '@/config/synapseConfigs/programs'
+import { CardContainerLogic } from 'synapse-react-client/components/CardContainerLogic/CardContainerLogic'
+
+function ExplorePrograms() {
+  return <CardContainerLogic {...programsCardContainerLogicProps} />
+}
+
+export default ExplorePrograms
