+++ apps/portals/ampals/src/pages/resources/PublishingRequirements.tsx
@@ -1,0 +1,13 @@
+import { MarkdownSynapse } from 'synapse-react-client/components/Markdown/MarkdownSynapse'
+
+function PublishingRequirements() {
+  return (
+    <MarkdownSynapse
+      ownerId="syn64892175"
+      wikiId="632170"
+      loadingSkeletonRowCount={20}
+    />
+  )
+}
+
+export default PublishingRequirements
